requirements.txt
amqp==2.4.1
asgiref==3.3.1
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.5
boto==2.46.1
boto3==1.4.4
botocore==1.5.16
celery==4.2.1
certifi==2019.6.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.7
cycler==0.10.0
dataclasses==0.6
decorator==4.0.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11.20
django-celery-beat==1.5.0
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-extensions==1.7.4
django-filter==2.0.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-serializable-model==0.0.3
django-storages==1.5.2
django-timezone-field==4.0
djangorestframework==3.9.4
docutils==0.13.1
enum34==1.1.6
future==0.14.3
futures==3.0.5
google-api-python-client==1.6.1
graphviz==0.10.1
gunicorn==19.4.5
httpie==0.9.6
httplib2==0.10.3
idna==2.6
ipython==5.1.0
ipython-genutils==0.1.0
jmespath==0.9.1
joblib==0.17.0
kombu==4.2.2.post1
llvmlite==0.31.0
Markdown==3.0.1
mysql-connector-python==8.0.22
nltk==3.0.2
nltk-tgrep==1.0.6
nr.types==1.1.1
numba==0.48.0
numpy==1.18.2
oauth2client==4.0.0
opencv-contrib-python==4.2.0.32
opencv-python==3.4.3.18
pandas==0.23.4
pathlib==1.0.1
pathlib2==2.1.0
pickleshare==0.7.4
Pillow==5.4.1
prompt-toolkit==1.0.7
psycopg2==2.8.2
pyasn1==0.2.1
pyasn1-modules==0.0.8
pycparser==2.19
Pygments==2.1.3
pyparsing==2.0.3
python-crontab==2.4.0
python-dateutil==2.8.1
python-decouple==3.1
pytz==2014.10
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.10
scandir==1.7
scipy==1.1.0
selenium==3.141.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
traitlets==4.3.0
typing==3.6.6
uritemplate==3.0.0
urllib3==1.22
vine==1.2.0
virtualenv==15.0.3
wcwidth==0.1.7
whitenoise==4.1.2
win-unicode-console==0.5
winshell==0.6
